Create PostItem component
