+++ client/constants/Post.ts
@@ -1,0 +1,6 @@
+interface Post {
+    category: 'Basic' | 'Intermediate' |  'Advanced';
+    slug: string;
+}
+
+export default Post;
